tests/test_strategies.py
"""
Unit tests for detection strategies
"""

import pytest
from datetime import datetime

# Import strategies when available
# from src.strategies.arbitrage import ArbitrageDetector
# from src.strategies.momentum import MomentumDetector
# from src.strategies.volume import VolumeAnalyzer
# from src.strategies.orderflow import OrderFlowAnalyzer


class TestArbitrageDetector:
    """Tests for arbitrage detection"""

    def test_detects_arbitrage_gap(self):
        """Should detect when YES + NO < 1.00"""
        # ArbitrageDetector with 2% threshold
        # yes_price = 0.45, no_price = 0.50 → total = 0.95 → 5% gap
        # Should trigger signal
        assert True  # Placeholder

    def test_ignores_efficient_market(self):
        """Should not trigger when market is efficient"""
        # yes_price = 0.50, no_price = 0.50 → total = 1.00 → no gap
        # Should not trigger
        assert True  # Placeholder

    def test_calculates_correct_gap(self):
        """Should calculate gap correctly in cents"""
        # yes = 0.45, no = 0.50 → gap = 5 cents
        assert True  # Placeholder


class TestMomentumDetector:
    """Tests for momentum detection"""

    def test_detects_bullish_momentum(self):
        """Should detect upward price movement"""
        # Feed prices: 0.50, 0.52, 0.55, 0.58
        # Should detect bullish momentum
        assert True  # Placeholder

    def test_detects_bearish_momentum(self):
        """Should detect downward price movement"""
        # Feed prices: 0.50, 0.48, 0.45, 0.42
        # Should detect bearish momentum
        assert True  # Placeholder

    def test_requires_minimum_history(self):
        """Should not trigger without sufficient history"""
        # Feed only 2 prices when min_history = 3
        # Should return None
        assert True  # Placeholder


class TestVolumeAnalyzer:
    """Tests for volume anomaly detection"""

    def test_detects_volume_spike(self):
        """Should detect when volume exceeds z-score threshold"""
        # Normal volumes: [1000, 1100, 900, 1050, ...]
        # Then: 5000 → z-score > 2.5 → should trigger
        assert True  # Placeholder

    def test_calculates_correct_z_score(self):
        """Should calculate z-score correctly"""
        # Given known mean and std, verify calculation
        assert True  # Placeholder


class TestOrderFlowAnalyzer:
    """Tests for order flow analysis"""

    def test_detects_buy_pressure(self):
        """Should detect heavy bid-side volume"""
        # bids = [(0.50, 1000), (0.49, 2000)]
        # asks = [(0.51, 500), (0.52, 300)]
        # ratio = 3000 / 800 = 3.75 > 3.0 → buy pressure
        assert True  # Placeholder

    def test_detects_sell_pressure(self):
        """Should detect heavy ask-side volume"""
        # bids = [(0.50, 300), (0.49, 200)]
        # asks = [(0.51, 1000), (0.52, 2000)]
        # ratio = 500 / 3000 = 0.17 < 0.33 → sell pressure
        assert True  # Placeholder


class TestRiskCalculator:
    """Tests for risk scoring"""

    def test_calculates_composite_score(self):
        """Should compute weighted risk score"""
        assert True  # Placeholder

    def test_assigns_correct_tier(self):
        """Should assign low/medium/high based on thresholds"""
        assert True  # Placeholder


if __name__ == "__main__":
    pytest.main([__file__, "-v"])
