requirements.txt
httpx[socks]>=0.27.0
tweepy>=4.14.0
python-dotenv>=1.0.1
Pillow>=10.4.0
aiohttp>=3.9.3
